+++ RELEASES.md
@@ -9,8 +9,10 @@ Version 0.5.3 (unreleased)
 ==========================
 
   * `Rem` and `RemAssign` were implemented for fixed-point numbers.
-  * The method `checked_rem` was added to all fixed-point types and to
-    the `Fixed` trait.
+  * The following methods were added to all fixed-point types and to
+    the `Fixed` trait:
+	  * `checked_rem`
+	  * `div_euclid`, `rem_euclid`x
 
 Version 0.5.2 (2020-02-02)
 ==========================